dialogue system revamped, but right now its behind overlay
