+++ Assets/Prefabs/UI/MessageUI.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1d0e708991ccf1940adda25817c5531f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
