+++ Assets/Scripts/Dialogue.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 019214678fca576439883a07c72a997e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
