+++ Assets/Scripts/Dialogue/DialogueScriptableObj.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "DialogueScriptableObj", menuName = "Scriptable Objects/DialogueScriptableObj")]
+public class DialogueScriptableObj : ScriptableObject
+{
+    public string[] lines;
+    public Sprite spirteA;
+    public Sprite spirteB;
+}
