+++ Assets/Scripts/Dialogue/DialogueScriptableObj.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: dd9abb2785530084480ddfaf79158313
