+++ Assets/Scripts/Dialogue/Tutorial 1.asset
@@ -1,0 +1,20 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: dd9abb2785530084480ddfaf79158313, type: 3}
+  m_Name: Tutorial 1
+  m_EditorClassIdentifier: Assembly-CSharp::DialogueScriptableObj
+  lines:
+  - According to our data
+  - This software contains an array of Earthlings we can meet and study
+  - Get to swiping!
+  spirteA: {fileID: 21300000, guid: 7cf1f3e381de4504d8c4445353b39e1d, type: 3}
+  spirteB: {fileID: 21300000, guid: 14195ce6032b9da4ab6a0eba52e7c89b, type: 3}
