+++ Assets/Scripts/Dialogue/Tutorial 1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dd9ec0c25f4b5434d807fd5409053d54
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
