+++ Assets/Scripts/Dialogue/Tutorial 2.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dcc8cb46290c3cc40846f21b1904c14f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
