--- Assets/Scripts/DialogueSystem.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a081637e13b561142aebde0e04b95161
