+++ Assets/Shaders/Camera Filters/Outside Filter.renderTexture
@@ -12,8 +12,8 @@ RenderTexture:
     Hash: 00000000000000000000000000000000
   m_IsAlphaChannelOptional: 0
   serializedVersion: 6
-  m_Width: 63
-  m_Height: 35
+  m_Width: 320
+  m_Height: 180
   m_AntiAliasing: 1
   m_MipCount: -1
   m_DepthStencilFormat: 94
