+++ Assets/Sprites/Dialogue Portrait.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 51c351457c46c3c4ab4fd2b67b9012a6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
